+++ code/src/itest/pcgen/io/BasicSaveRestoreTest.java
@@ -29,7 +29,7 @@ import pcgen.core.PCTemplate;
 import pcgen.core.SettingsHandler;
 import pcgen.core.Skill;
 import pcgen.io.testsupport.AbstractSaveRestoreTest;
-import pcgen.output.channel.ChannelCompatibility;
+import pcgen.output.channel.compat.AlignmentCompat;
 import plugin.lsttokens.pcclass.HdToken;
 
 public class BasicSaveRestoreTest extends AbstractSaveRestoreTest
@@ -67,7 +67,7 @@ public class BasicSaveRestoreTest extends AbstractSaveRestoreTest
 	public void testAlignment()
 	{
 		finishLoad();
-		ChannelCompatibility.setCurrentAlignment(pc.getCharID(), le);
+		AlignmentCompat.setCurrentAlignment(pc.getCharID(), le);
 		runRoundRobin(null);
 	}
