+++ code/src/itest/tokenmodel/testsupport/AbstractGrantedListTokenTest.java
@@ -27,7 +27,7 @@ import pcgen.core.EquipmentModifier;
 import pcgen.core.PCCheck;
 import pcgen.core.PCStat;
 import pcgen.core.character.CompanionMod;
-import pcgen.output.channel.ChannelCompatibility;
+import pcgen.output.channel.compat.AlignmentCompat;
 
 public abstract class AbstractGrantedListTokenTest<T extends CDOMObject>
 		extends AbstractAddListTokenTest<T>
@@ -38,10 +38,10 @@ public abstract class AbstractGrantedListTokenTest<T extends CDOMObject>
 		T granted = createGrantedObject();
 		processToken(lg);
 		assertEquals(0, getCount());
-		ChannelCompatibility.setCurrentAlignment(pc.getCharID(), lg);
+		AlignmentCompat.setCurrentAlignment(pc.getCharID(), lg);
 		assertTrue(containsExpected(granted));
 		assertEquals(1, getCount());
-		ChannelCompatibility.setCurrentAlignment(pc.getCharID(), ng);
+		AlignmentCompat.setCurrentAlignment(pc.getCharID(), ng);
 		assertEquals(0, getCount());
 		assertTrue(cleanedSideEffects());
 	}
