+++ code/src/java/pcgen/core/npcgen/NPCGenerator.java
@@ -57,7 +57,7 @@ import pcgen.core.character.CharacterSpell;
 import pcgen.core.pclevelinfo.PCLevelInfo;
 import pcgen.core.spell.Spell;
 import pcgen.gui2.UIPropertyContext;
-import pcgen.output.channel.ChannelCompatibility;
+import pcgen.output.channel.compat.AlignmentCompat;
 import pcgen.util.Logging;
 import pcgen.util.chooser.ChooserFactory;
 import pcgen.util.enumeration.Visibility;
@@ -549,7 +549,7 @@ public final class NPCGenerator
 				{
 					Logging.debugPrint("NPCGenerator: Selected " + randAlign //$NON-NLS-1$
 						+ " for alignment " + align); //$NON-NLS-1$
-					ChannelCompatibility.setCurrentAlignment(aPC.getCharID(), randAlign);
+					AlignmentCompat.setCurrentAlignment(aPC.getCharID(), randAlign);
 				}
 
 				final Race r = getRace(aRace);
