+++ code/src/java/pcgen/output/channel/ChannelCompatibility.java
@@ -21,7 +21,6 @@ import pcgen.cdom.enumeration.CharID;
 import pcgen.cdom.util.CControl;
 import pcgen.cdom.util.ControlUtilities;
 import pcgen.core.Globals;
-import pcgen.core.PCAlignment;
 import pcgen.core.PCStat;
 import pcgen.facade.util.WriteableReferenceFacade;
 import pcgen.output.channel.compat.StatAdapter;
@@ -63,33 +62,5 @@ public final class ChannelCompatibility
 		{
 			return (WriteableReferenceFacade<Number>) context.getVariableContext().getChannel(id, stat, channelName);
 		}
-	}
-
-	/**
-	 * Gets the current Alignment for the PC represented by the given CharID.
-	 * 
-	 * @param id
-	 *            The CharID representing the PC for which the Alignment should be
-	 *            returned
-	 * @return The current Alignment for the PC represented by the given CharID
-	 */
-	public static PCAlignment getCurrentAlignment(CharID id)
-	{
-		String channelName = ControlUtilities.getControlToken(Globals.getContext(), CControl.ALIGNMENTINPUT);
-		return (PCAlignment) ChannelUtilities.readGlobalChannel(id, channelName);
-	}
-
-	/**
-	 * Sets the current alignment for the PC represented by the given CharID.
-	 * 
-	 * @param id
-	 *            The CharID representing the PC for which the Alignment should be set
-	 * @param align
-	 *            The Alignment which should be set
-	 */
-	public static void setCurrentAlignment(CharID id, PCAlignment align)
-	{
-		String channelName = ControlUtilities.getControlToken(Globals.getContext(), CControl.ALIGNMENTINPUT);
-		ChannelUtilities.setGlobalChannel(id, channelName, align);
 	}
 }
