+++ code/src/test/pcgen/core/PrereqHandlerTest.java
@@ -4,7 +4,7 @@ import pcgen.AbstractCharacterTestCase;
 import pcgen.cdom.enumeration.StringKey;
 import pcgen.core.prereq.PrereqHandler;
 import pcgen.core.prereq.Prerequisite;
-import pcgen.output.channel.ChannelCompatibility;
+import pcgen.output.channel.compat.AlignmentCompat;
 import pcgen.persistence.PersistenceLayerException;
 import pcgen.persistence.lst.prereq.PreParserFactory;
 import plugin.lsttokens.testsupport.BuildUtilities;
@@ -43,13 +43,13 @@ public class PrereqHandlerTest extends AbstractCharacterTestCase
 		human.setName("Human");
 		pc.setRace(human);
 
-		ChannelCompatibility.setCurrentAlignment(pc.getCharID(), le);
+		AlignmentCompat.setCurrentAlignment(pc.getCharID(), le);
 		assertEquals("Non-negate returns false", false, PrereqHandler.passes(
 			prereq, pc, null));
 		assertEquals("Negate returns false", false, PrereqHandler.passes(
 			prereqNeg, pc, null));
 
-		ChannelCompatibility.setCurrentAlignment(pc.getCharID(), tn);
+		AlignmentCompat.setCurrentAlignment(pc.getCharID(), tn);
 		assertEquals("Non-negate returns true", true, PrereqHandler.passes(
 			prereq, pc, null));
 		assertEquals("Negate returns true", true, PrereqHandler.passes(
